detect/uri: Move uri logic into a separate module
