+++ rust/src/detect/mod.rs
@@ -17,6 +17,7 @@
 
 pub mod byte_math;
 pub mod uint;
+pub mod uri;
 pub mod detect;
 pub mod error;
 pub mod parser;
